+++ src/stamina/_config.py
@@ -4,29 +4,32 @@
 
 from __future__ import annotations
 
+from dataclasses import dataclass
+from threading import Lock
+from typing import Iterable
 
-class _Active:
-    _is_active: bool
+from ._instrumentation import INSTRUMENTS
+from .typing import RetryHook
 
 
-_ACTIVE = _Active()
-_ACTIVE._is_active = True
+@dataclass
+class Config:
+    is_active: bool
+    on_retry: Iterable[RetryHook]
+
+
+_CONFIG = Config(is_active=True, on_retry=INSTRUMENTS)
+_LOCK = Lock()
 
 
 def is_active() -> bool:
     """
     Check whether retrying is active.
 
-    .. warning::
-
-        Please note that these APIs are **not** thread-safe and are **not**
-        meant to be used in regular program flow. Their purposse is to speed up
-        test suites.
-
     Returns:
         Whether retrying is active.
     """
-    return _ACTIVE._is_active
+    return _CONFIG.is_active
 
 
 def set_active(active: bool) -> None:
@@ -35,4 +38,5 @@ def set_active(active: bool) -> None:
 
     Is idempotent and can be called repeatedly with the same value.
     """
-    _ACTIVE._is_active = bool(active)
+    with _LOCK:
+        _CONFIG.is_active = bool(active)