+++ src/stamina/_sync.py
@@ -10,9 +10,10 @@ from collections.abc import Callable
 from functools import wraps
 from typing import TypeVar
 
-import tenacity
+import tenacity as _t
 
-from . import _control
+from ._config import _CONFIG
+from ._instrumentation import guess_name
 
 
 if sys.version_info >= (3, 10):
@@ -41,8 +42,7 @@ def retry(
     """
     Retry if one of configured exceptions are raised.
 
-    The back-off delays between retries grow exponentially plus a random
-    jitter.
+    The backoff delays between retries grow exponentially plus a random jitter.
 
     Example:
 
@@ -52,15 +52,15 @@ def retry(
 
             from stamina import retry
 
-            @retry(on=httpx.HTTPError, attempts=3) def do_it(code: int) ->
-            httpx.Response:
+            @retry(on=httpx.HTTPError, attempts=3)
+            def do_it(code: int) -> httpx.Response:
                 resp = httpx.get(f"https://httpbin.org/status/{code}")
                 resp.raise_for_status()
 
                 return resp
     """
-    retry_ = tenacity.retry_if_exception_type(on)
-    wait = tenacity.wait_exponential_jitter(
+    retry_ = _t.retry_if_exception_type(on)
+    wait = _t.wait_exponential_jitter(
         initial=wait_initial,
         max=wait_max,
         exp_base=wait_exp_base,
@@ -69,13 +69,40 @@ def retry(
     stop = _make_stop(attempts=attempts, timeout=timeout)
 
     def retry_decorator(wrapped: Callable[P, T]) -> Callable[P, T]:
+        name = guess_name(wrapped)
+
         @wraps(wrapped)
         def inner(*args: P.args, **kw: P.kwargs) -> T:
-            if not _control._ACTIVE._is_active:
+            if not _CONFIG.is_active:
                 return wrapped(*args, **kw)
 
-            for attempt in tenacity.Retrying(
-                retry=retry_, wait=wait, stop=stop, reraise=True
+            before_sleep: Callable[[_t.RetryCallState], None] | None
+            if _CONFIG.on_retry:
+
+                def before_sleep(rcs: _t.RetryCallState) -> None:
+                    attempt = rcs.attempt_number
+                    exc = rcs.outcome.exception()
+                    backoff = rcs.idle_for
+
+                    for hook in _CONFIG.on_retry:
+                        hook(
+                            attempt=attempt,
+                            backoff=backoff,
+                            exc=exc,
+                            name=name,
+                            args=args,
+                            kwargs=kw,
+                        )
+
+            else:
+                before_sleep = None
+
+            for attempt in _t.Retrying(
+                retry=retry_,
+                wait=wait,
+                stop=stop,
+                reraise=True,
+                before_sleep=before_sleep,
             ):
                 with attempt:
                     res = wrapped(*args, **kw)
@@ -87,23 +114,21 @@ def retry(
     return retry_decorator
 
 
-def _make_stop(
-    *, attempts: int | None, timeout: float | None
-) -> tenacity.stop_base:
+def _make_stop(*, attempts: int | None, timeout: float | None) -> _t.stop_base:
     """
     Combine *attempts* and *timeout* into one stop condition.
     """
     stops = []
 
     if attempts:
-        stops.append(tenacity.stop_after_attempt(attempts))
+        stops.append(_t.stop_after_attempt(attempts))
     if timeout:
-        stops.append(tenacity.stop_after_delay(timeout))
-
-    if not stops:
-        return tenacity.stop_never
+        stops.append(_t.stop_after_delay(timeout))
 
     if len(stops) > 1:
-        return tenacity.stop_any(*stops)
+        return _t.stop_any(*stops)
+
+    if not stops:
+        return _t.stop_never
 
     return stops[0]